Fix linters errors
