+++ src/taskClass.js
@@ -1,0 +1,7 @@
+export default class Task {
+  constructor(index, description) {
+    this.index = index;
+    this.description = description;
+    this.completed = false;
+  }
+}
